adds "add slot" button to calendar
Adds a "adds slot" button to calendar that triggers the new slot modal setting the time to the next hour based on current time and a default duration of 1 hour.
Adds a datepicker to modal's step 3 so admin can choose a different day.
Enables dayClick event on calendar that triggers the new slot modal setting the time where the calendar was clicked and a default duration of 1 hour